puntos en reportes

+++ PRIII-24-ESCUELA-PROGRAMACION-API/Models/Competencia.cs
@@ -2,10 +2,10 @@
 {
     public class Competencia
     {
-        public int Id { get; set; }
+        public UInt32 Id { get; set; }
         public string Titulo { get; set; }
         public string Descripcion { get; set; } 
-        public int Puntos { get; set; }
+        public UInt16 Puntos { get; set; }
         public DateTime Fecha_Inicio { get; set; }
         public DateTime Fecha_Fin { get; set; }
         public char Estado { get; set; }
